// src/shared/types/database.ts
export interface Word {
  id: number;
  text_uthmani: string;
  text_simple: string;
  translation_bm: string | null;
  translation_en: string | null;
  transliteration: string | null;
  root: string | null;
  lemma: string | null;
  pos: string | null;
  frequency: number;
}